requirements.txt
PyYAML==5.3.1
joblib==0.14.1
dagshub==0.0.2rc1
matplotlib
plac==1.1.3
pandas==1.0.3
pytest==5.4.1
dvc==0.93.0
numpy==1.18.3
scikit-learn==0.22.2.post1
